csma: Add accessors/mutators for channel fields
Rather than accessing m_state, m_currentSrc, and m_currentPkt directly,
these are now handled via private accessors and mutators, which also
take the deviceId as a parameter (similar to the other extended APIs
from the previous commit).

Again, this is not currently useful, but will be used to support full
duplex operation.