gtk-config-store: Sort class names
There are a lot of classes with and it's difficult to find the one you're
looking for when they're not in any order.  This sorts them, but only in
the defaults list.  It'd be nice to do the same for the objects list too,
but it's not as pressing a need since there are generally many fewer.