First shot at adding ECMP to Ipv4StaticRouting
This adds ECMP to the IPv4 static routing code.  It is mostly just a
remix of the ECMP code that has been added to (or proposed for) IPv4
global routing by Wilson Thong and Tom Henderson (see Bug 667).  Much
of it could probably be refactored (especially true in the case of
GetTupleValue() (here renamed HashHeaders()) for which there need be
only a single implementation).